# 2x faster MosesTokenizer